// packages/chromator/src/enums/ColourCodeFormat.ts
export enum ColourCodeFormat {
  HslObject = 'HslObject',
  HslString = 'HslString',
  HslaObject = 'HslaObject',
  HslaString = 'HslaString',
  HsvObject = 'HsvObject',
  HsvaObject = 'HsvaObject',
  NamedColour = 'NamedColour',
  NamedTransparent = 'NamedTransparent',
  Rgb255Object = 'RgbObject',
  Rgb255String = 'Rgb255String',
  RgbHexLongWithHash = 'RgbHexLongWithHash',
  RgbHexLongWithoutHash = 'RgbHexLongWithoutHash',
  RgbHexShortWithHash = 'RgbHexShortWithHash',
  RgbHexShortWithoutHash = 'RgbHexShortWithoutHash',
  Rgba255Object = 'RgbaObject',
  Rgba255String = 'Rgba255String',
  RgbaHexLongWithHash = 'RgbaHexLongWithHash',
  RgbaHexLongWithoutHash = 'RgbaHexLongWithoutHash',
  RgbaHexShortWithHash = 'RgbaHexShortWithHash',
  RgbaHexShortWithoutHash = 'RgbaHexShortWithoutHash',
}